requirements.txt
torch
torchvision
numpy
tqdm
gnutools-python
requests
pyYaml
